requirements.txt
Flask
tensorflow
numpy
Pillow
requests
python-dotenv
langchain
langchain-community
faiss-cpu
pypdf
docx2txt
langchain-huggingface
sentence-transformers
